% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{eff.arm.infofract}
\alias{eff.arm.infofract}
\title{information-fraction based arm efficacy stop}
\usage{
eff.arm.infofract(posterior, b, n, N, p)
}
\arguments{
\item{posterior}{posterior probability of P(theta>delta) > b}

\item{b}{probability boundary for efficacy}

\item{n}{number of recruited participants}

\item{N}{total (planned) sample size}

\item{p}{tuning parameter}
}
\value{
\link{eff.arm.infofract} returns a logical constant.
}
\description{
allows stopping an arm for efficacy at a given look when the probability of the corresponding target parameter being greater than delta is greater than a function of the information fraction at that look.
}
