% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{eff.arm.simple}
\alias{eff.arm.simple}
\title{Simple arm efficacy stop}
\usage{
eff.arm.simple(posterior, b)
}
\arguments{
\item{posterior}{posterior probability of P(theta>delta) > b}

\item{b}{probability boundary for efficacy}
}
\value{
\link{eff.arm.simple} returns a logical constant.
}
\description{
allows stopping an arm for efficacy when the probability of the corresponding target parameter being greater than delta is greater than a fixed value b
}
