% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{eff.trial.all}
\alias{eff.trial.all}
\title{trial efficacy stop}
\usage{
eff.trial.all(eff.target)
}
\arguments{
\item{eff.target}{logical vector with length of active intervention arms}
}
\value{
\link{eff.trial.all} returns a logical constant.
}
\description{
allows stopping the trial for efficacy at a given look if all active treatments reached efficacy
}
