% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{eff.trial.any}
\alias{eff.trial.any}
\title{trial efficacy stop}
\usage{
eff.trial.any(eff.target)
}
\arguments{
\item{eff.target}{logical vector with length of active intervention arms}
}
\value{
\link{eff.trial.any} returns a logical constant.
}
\description{
allows stopping the trial for efficacy at a given look if at least one active treatment reached efficacy
}
