% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{fut.arm.simple}
\alias{fut.arm.simple}
\title{arm futility stop}
\usage{
fut.arm.simple(posterior, b)
}
\arguments{
\item{posterior}{the BATS ingredient '\code{posterior}' corresponding, in this context, to the (posterior) probability P(beta>delta.fut) for a given target parameter.}

\item{b}{the cut-off value to declare futility (to be defined in \code{fut.arm.control}).}
}
\value{
\link{fut.arm.simple} returns a logical constant.
}
\description{
allows stopping an arm for futility when the probability of the corresponding target parameter being greater than delta is smaller than a fixed value 'b'
}
