% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{fut.trial.all}
\alias{fut.trial.all}
\title{trial futility stop}
\usage{
fut.trial.all(fut.target)
}
\arguments{
\item{fut.target}{the 'BATSS' ingredient '\code{fut.target}' corresponding to a \link[base]{logical} vector of the same length as argument \code{which} (i.e., the number of target parameters) indicating if futility was declared for each target parameter at that stage or at a previous stage.}
}
\value{
\link{fut.trial.all} returns a logical constant.
}
\description{
allows stopping the trial for efficacy if \emph{all} active treatment reached futility at the look of interest or before.
}
