% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bats.R
\name{plot.bats}
\alias{plot.bats}
\title{Plot function for BATS outputs}
\usage{
\method{plot}{bats}(x, focus = "size", hypothesis = "H1", ...)
}
\arguments{
\item{x}{An object of class 'bats' (i.e., output of the function BATS::bats).}

\item{focus}{A character string indicating the type of plot with options 'size' (default) to display the total and per group sample size observed in the Monte Carlo trials, and 'estimates' to display the Monte Carlo trial target estimates as a function of the sample size.}

\item{hypothesis}{A character string indicating which alternative hypothesis to use for analyses considering both "H0" and "H1", with options "H1" (default) and "H0".}

\item{...}{Additional arguments affecting the plot produced.}
}
\description{
Plot for objects of class 'bats'
}
