% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_generic.R
\name{print.bats}
\alias{print.bats}
\title{Print function for BATS outputs}
\usage{
\method{print}{bats}(x, ...)
}
\arguments{
\item{x}{An object of class 'bats'.}

\item{...}{Additional arguments affecting the print produced.}
}
\description{
Print method function for objects of class 'bats' (i.e., output of the function \link{bats.glm}).
}
\seealso{
\code{\link[=bats.glm]{bats.glm()}}, the function generating S3 objects of class 'bats'.
}
