% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_generic.R
\name{print.batss}
\alias{print.batss}
\title{Print function for BATSS outputs}
\usage{
\method{print}{batss}(x, ...)
}
\arguments{
\item{x}{An object of class 'batss'.}

\item{...}{Additional arguments affecting the print produced.}
}
\description{
Print method function for objects of class 'batss' (i.e., output of the function \link{batss.glm}).
}
\seealso{
\code{\link[=batss.glm]{batss.glm()}}, the function generating S3 objects of class 'batss'.
}
