% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{RAR.optimal}
\alias{RAR.optimal}
\title{'Optimal' control allocation}
\usage{
RAR.optimal(active)
}
\arguments{
\item{active}{the BATS ingredient '\code{active}' corresponding to a \link[base]{logical} vector of the same length as argument \code{which} (i.e., the number of target parameters) indicating if each arm is active at the look of interest.}
}
\value{
\link{RAR.optimal} returns a vector of probabilities with length of active.
}
\description{
technically not response adaptive but keeps allocation ratio to control at the square root of active intervention arms
}
