% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{RAR.trippa}
\alias{RAR.trippa}
\title{RAR of Trippa et al. (2012)}
\usage{
RAR.trippa(posterior, n, N, ref, active, gamma, eta, nu)
}
\arguments{
\item{posterior}{posterior probability of P(theta>delta) < b}

\item{n}{current sample size}

\item{N}{total sample size}

\item{ref}{vector indicating the reference group}

\item{active}{vector indicating which treatments are active}

\item{gamma}{scaling factor}

\item{eta}{scaling factor}

\item{nu}{scaling factor}
}
\value{
\link{RAR.trippa} returns a vector of probabilities with length of active.
}
\description{
define the group allocation probabilities based on the response adaptive randomisation rule of Trippa et al. (2012)
}
