% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bats.R
\name{summary.bats}
\alias{summary.bats}
\title{Summary function for BATS outputs}
\usage{
\method{summary}{bats}(object, full = FALSE, ...)
}
\arguments{
\item{object}{An object of class 'bats' (i.e., output of the function BATS::bats).}

\item{full}{A logical indicating if a standard (full = FALSE, default) or extended output (full = TRUE) should be returned.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Summary method function for objects of class 'bats'.
}
\examples{
\dontrun{
data(sim2B)
plot(sim2B,focus="size",hypothesis="H0")
plot(sim2B,focus="size",hypothesis="H1")
plot(sim2B,focus="estimates",hypothesis="H0")
plot(sim2B,focus="estimates",hypothesis="H1")
}
}
