% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{alloc.simple}
\alias{alloc.simple}
\title{Simple allocation function}
\usage{
alloc.simple(m, prob)
}
\arguments{
\item{m}{scalar of number of participants to be allocated}

\item{prob}{named vector of allocation ratio or probabilities}
}
\value{
\link{alloc.simple} returns an object of class \link[base]{factor} with levels as provided in the input vector 'prob' and length 'm'.
}
\description{
\link{alloc.simple} independently randomises each unit to a group (i.e., flips a coin for each unit) so that the observed allocation probabilities may be far from the target ones.
}
\examples{
alloc.simple(100, prob = c(A=.4,B=.6))
table(alloc.simple(100, prob = c(A=.4,B=.6)))
table(alloc.simple(100, prob = c(A=.4,B=.6)))
}
