% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_combine.R
\name{bats.combine}
\alias{bats.combine}
\title{Combines BATS outputs}
\usage{
bats.combine(paths)
}
\arguments{
\item{paths}{Vector indicating the paths to the rdata files containing the outputs of the function BATS::bats.glm considering the same trial design but different set of seeds. This requires the argument 'full' of the function BATS::bats.glm to be 'TRUE'.}
}
\description{
Combines different evaluations of BATS::bats.glm considering the same trial design but different sets of seeds. This function is useful when the evaluation of Monte Carlo samples generated by different seeds was split in sets computed by different nodes/cpus. The output of this function is of class 'bats' meaning that the usual generic functions (print, summary, plot) can be used.
}
\examples{
\dontrun{
data(sim2B)
plot(sim2B,focus="size",hypothesis="H0")
plot(sim2B,focus="size",hypothesis="H1")
plot(sim2B,focus="estimates",hypothesis="H0")
plot(sim2B,focus="estimates",hypothesis="H1")
}
}
