% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{batss.surv}
\alias{batss.surv}
\title{Bayesian adaptive trial simulations for time-to-event data}
\usage{
batss.surv(
  model,
  family = "weibullsurv",
  surv = simsurv::simsurv,
  surv.control,
  var,
  var.control = NULL,
  fup = NULL,
  cens = NULL,
  cens.control = NULL,
  accr = runif,
  accr.control = list(min = 0, max = 3),
  accr.type = "fixed",
  hr,
  which,
  R = 10000,
  N,
  n.max = NULL,
  alternative = "less",
  RAR = NULL,
  RAR.control = NULL,
  interim,
  t.max = NULL,
  event.max = NULL,
  prob0,
  delta.eff = 0,
  delta.fut = delta.eff,
  delta.RAR = 0,
  eff.arm,
  eff.trial = NULL,
  eff.arm.control = NULL,
  eff.trial.control = NULL,
  fut.arm,
  fut.trial = NULL,
  fut.arm.control = NULL,
  fut.trial.control = NULL,
  H0 = TRUE,
  computation = "parallel",
  mc.cores = getOption("mc.cores", 3L),
  extended = 0,
  ...
)
}
\arguments{
\item{model}{an object of class '\link[stats]{formula}' indicating a symbolic description of the model to be fitted}

\item{family}{A character string indicating the log-survival likelihood as described in the package INLA. Default set to '\code{weibullsurv}', other options include '\code{exponential.surv}', '\code{loglogistic.surv}','\code{lognormal.surv}', '\code{coxph}'. See \link[INLA]{inla.list.models}.}

\item{surv}{Time-to-event data simulator function. Currently, '\link[simsurv]{simsurv}' is the only option and used by default.}

\item{surv.control}{A list of parameters for the call to '\link[simsurv]{simsurv}'.}

\item{var}{A list. Each entry corresponds to a variable described under '\code{model}' and indicates the name of a function allowing to generate variates (like \link[stats]{rnorm} and \link[stats]{rexp}, for example). The list names have to match the variable names used in '\code{model}'. The grouping variable corresponding to the target parameters has to be of class '\link[base]{factor}' with levels corresponding to the names indicated in argument \code{prob0} (see below).}

\item{var.control}{An optional list of control parameters for the functions indicated in '\code{var}'. The names of the list items need to correspond to the names used in '\code{var}'. Each element is another list with names of the elements corresponding to the parameter names of the functions specified in '\code{var}'.}

\item{fup}{A scalar specifying the follow-up time of the last patient recruited and hence the final analysis, this only applies if 'maxt=NULL' in 'surv.control'.}

\item{cens}{Random number generating function for censoring, e.g. '\link[stats]{rexp}'}

\item{cens.control}{Additional parameters for call to censoring function.}

\item{accr}{Random number generating function for entry times, e.g. '\link[stats]{rexp}'}

\item{accr.control}{Additional parameters for call to entry times generating function.}

\item{accr.type}{One of 'random' or 'fixed', the former will draw entry times from the distribution defined in '\code{accr}', adding the resulting times results in a random study duration. Setting this parameter to 'fixed' will allow setting a fixed study length and '\code{accr}' will be used to describe the distribution within the interval from '\code{0}' to study duration.}

\item{hr}{A numerical vector of parameter values for the hazard ratios. Its length has to match the number of treatments (excluding control).}

\item{which}{A numerical vector indicating the position of the target parameters.}

\item{R}{a vector of natural numbers to be used as seeds (check \link[base]{set.seed}) for the different Monte Carlo trials (the vector length will thus correspond to the number of Monte Carlo trials). When \code{R} is a scalar, seeds \code{1} to \code{R} are used, where \code{R} corresponds to the number of Monte Carlo trials.}

\item{N}{A scalar indicating the total sample size.}

\item{n.max}{A vector indicating the maximum sample size per arm, if a scalar is given the same size will be used for all arms.}

\item{alternative}{A vector of strings providing the one-sided direction of the alternative hypothesis corresponding to each target parameter indicated under '\code{which}' (in the same order). Possibilities are 'greater' or 'less'. If the vector is of length 1, the same direction will be used for all target parameter tests.}

\item{RAR}{A character string indicating the how the response-adaptive randomised probabilities corresponding to each group, reference group included, are defined. If RAR=NULL, equal probabilities of being attributed to each (active) group are used. Check ?BATS::prob.fun for examples.}

\item{RAR.control}{An optional list of control parameters the function in 'RAR'.}

\item{interim}{A list of parameters related to interim analyses. Possible list items include, '\code{recruited}' a vector of integers indicating the number of recruited participants at each look, last excluded, in increasing order, '\code{time}' a vector of integers indicating the time points of interim analyses, and '\code{event}' a vector of integers indicating the number of events. '\code{event.type}' is a character string describing the groups that are counted towards the numbers in '\code{event}' with options options '\code{all}', counting all events,'\code{control}', counting events in the control group only, '\code{cplusone}', counting the control plus treatments individually (this will result in different time points for the interims of the treatment groups), and '\code{min}', counting the last group to reach the threshold. See details for explanations and combinations.}

\item{t.max}{A scalar of the maximum trial duration, safeguard if 'events' takes too long to reach}

\item{event.max}{A scalar indicating the maximum number of event.max in the control group, once reached the trial will stop recruiting}

\item{prob0}{A named vector with initial allocation probabilities. Names need to correspond to the levels of the grouping variable. If \code{RAR = NULL}, these probabilities/ratios will be used throughout (fixed allocation probabilities).}

\item{delta.eff}{A vector (of length equal to the number of looks (i.e., number of interims + 1)) of clinically meaningful treatment effect values (on the linear predictor scale) to be used to define the efficacy-related posterior probabilities for each target parameter at each look. If a scalar is provided, the same value is used at each look. The default is \code{delta.eff = 0}.}

\item{delta.fut}{A vector (of length equal to the number of looks (i.e., number of interims + 1)) of clinically meaningful treatment effect values (on the linear predictor scale) to be used to define the futility-related posterior probabilities for each target parameter at each look. If a scalar is provided, the same value is used at each look. The default is \code{delta.fut = delta.eff}.}

\item{delta.RAR}{A vector (of length equal to the number of looks (i.e., number of interims + 1)) of clinically meaningful treatment effect values (on the linear predictor scale) to be used to define the RAR-related posterior probabilities for each target parameter at each look. If a scalar is provided, the same value is used at each interim analysis. The default is \code{delta.RAR = 0}. Note that, when a vector is provided, its last value is ignored as no randomisation is made at the last look.}

\item{eff.arm}{A function defining if efficacy has been achieved at a given look given the information available at that stage a given target parameter. The output of this function must be a \link[base]{logical} (of length 1). Arguments of this function will typically consider 'BATSS' ingredients. Check \link{eff.arm.simple} and \link{eff.arm.infofract} for examples.}

\item{eff.trial}{A function defining if the trial can be stopped for efficacy given the output of the function indicated in '\code{eff.arm}'. The output of this function must be a \link[base]{logical} of length one. Arguments of this function will typically only consider the 'BATSS' ingredient \code{eff.target}. Check \link{eff.trial.all} and \link{eff.trial.any} for examples. When \code{eff.trial = NULL} (default), the trial stops for efficacy when \emph{all} target parameters are found to be effective (like in \link{eff.trial.all}).}

\item{eff.arm.control}{An optional list of parameters for the function indicated in '\code{eff.arm}'.}

\item{eff.trial.control}{An optional list of parameters for the function indicated in '\code{eff.trial}'.}

\item{fut.arm}{A function defining if futility has been achieved at a given look given the information available at that stage for each target parameter. The output of this function must be a \link[base]{logical} (of length 1). Arguments of this function will typically consider 'BATSS' ingredients. Check \link{fut.arm.simple} to see an example of such a function.}

\item{fut.trial}{A function defining if the trial can be stopped for futility given the output of the function indicated in '\code{fut.arm}'. The output of this function must be a \link[base]{logical} of length one. Arguments of this function will typically only consider the 'BATSS' ingredient \code{fut.target}. Check \link{fut.trial.all} for an example of such a function. When \code{fut.trial = NULL} (default), the trial stops for futility when \emph{all} target parameters are found to be futile (like in \link{fut.trial.all}).}

\item{fut.arm.control}{An optional list of parameters for the function indicated in '\code{fut.arm}'.}

\item{fut.trial.control}{An optional list of parameters for the function indicated in '\code{fut.trial}'.}

\item{H0}{A logical indicating whether the simulation should also consider the case with all target parameters set to 0 to check the probability of rejecting the hypothesis that the target parameter value is equal to 0 individually (pairwise type I error) or globally (family-wise error rate). Default set to \code{H0=TRUE}.}

\item{computation}{A character string indicating how the computation should be performed. Possibilities are '\code{parallel}' or '\code{sequential}' with default \code{computation="parallel"} meaning that the computation is split between \code{mc.cores}.}

\item{mc.cores}{An integer indicating the number of CPUs to be used when \code{computation="parallel"} (Default to 3 if no global '\code{mc.cores}' global option is available via \link[base]{getOption}).}

\item{extended}{an integer indicating the type of results to be returned. 0 (default) provides summary statistics, 1 adds the results of each Monte Carlo trial and 2 additionally returns each Monte Carlo dataset. \link{batss.combine} requires extended > 0 as the function needs to merge results of different sets of seeds.}

\item{...}{Additional arguments to control fitting in \link[INLA]{inla}.}
}
\value{
The function \link{batss.surv} returns an S3 object of class 'batss' with available print/summary/plot functions
\itemize{
\item beta - A data frame providing information related to the beta parameter vector, like parameter names and values, for example.
\item look - A data frame providing information related to looks, like sample size of a given interim (m) and cumulative sample size at a given interim (n), for example.
\item par - A list providing different information, like the used seeds (seed) and the groups (group), for example.
\item H1 - A list providing trial results under the alternative, like the estimates per target parameter when the corresponding arm was stopped (estimate), the efficacy and futility probabilites per target parameter and overall (target, efficacy and futility), the sample size per group and trial (sample), the probabilities associated to each combination of efficacy and futility per group (scenario), the detailed results per trial (trial), for example.
\item H0 - A list providing trial results under the global null hypothesis (same structure as H1).
\item call - The matched call.
\item type - The type of 'BATSS' analysis, i.e. '\code{surv}'
}
}
\description{
Simulation of Bayesian adaptive trials with time-to-event endpoint using Integrated Nested Laplace Approximation (INLA).
}
\examples{
# Example 1: TBC
}
\seealso{
\link{summary.batss} and \link{plot.batss} for detailed summaries and plots, and \link{batss.combine} to combine different evaluations of \link{batss.surv} considering the same trial design but different sets of seeds (useful for cluster computation).
}
