% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{eff.arm.infofract}
\alias{eff.arm.infofract}
\title{information-fraction based arm efficacy stop}
\usage{
eff.arm.infofract(posterior, b, n, N, p)
}
\arguments{
\item{posterior}{the BATS ingredient '\code{posterior}' corresponding, in this context, to the (posterior) probability P(beta>delta.eff) for a given target parameter.}

\item{b}{a tuning parameter (to be defined in \code{eff.arm.control}).}

\item{n}{the BATS ingredient '\code{n}' corresponding to the vector of number of recruited participants per arm including the control group.}

\item{N}{the BATS ingredient '`N' corresponding to the maximum (planned) sample size.}

\item{p}{a tuning parameter (to be defined in \code{eff.arm.control}).}
}
\value{
\link{eff.arm.infofract} returns a logical constant.
}
\description{
allows stopping an arm for efficacy at a given look when the probability of the corresponding target parameter being greater than delta is greater than a function of the information fraction at that look.
}
