% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{eff.arm.simple}
\alias{eff.arm.simple}
\title{Simple arm efficacy stop}
\usage{
eff.arm.simple(posterior, b)
}
\arguments{
\item{posterior}{the BATS ingredient '\code{posterior}' corresponding, in this context, to the (posterior) probability of the target parameter being greater or smaller (depending on the argument \code{'alternative'} of \link{bats.glm}) than '\code{delta.eff}'.}

\item{b}{the cut-off value to declare efficacy.}
}
\value{
\link{eff.arm.simple} returns a logical constant.
}
\description{
allows stopping an arm for efficacy when the probability of the corresponding target parameter being greater than \code{delta.eff} is greater than a fixed value \code{b}.
}
