% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{eff.trial.all}
\alias{eff.trial.all}
\title{trial efficacy stop}
\usage{
eff.trial.all(eff.target)
}
\arguments{
\item{eff.target}{the 'BATSS' ingredient '\code{eff.target}' corresponding to a \link[base]{logical} vector of the same length as argument \code{which} (i.e., the number of target parameters) indicating if efficacy was reached for each target parameter at that stage or at a previous stage.}
}
\value{
\link{eff.trial.all} returns a logical constant.
}
\description{
allows stopping the trial for efficacy if \emph{all} target parameters reached efficacy at the look of interest or before.
}
