% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{eff.trial.any}
\alias{eff.trial.any}
\title{trial efficacy stop}
\usage{
eff.trial.any(eff.target)
}
\arguments{
\item{eff.target}{the BATSS ingredient '\code{eff.target}' corresponding to a \link[base]{logical} vector of the same length as argument \code{which} (i.e., the number of target parameters) indicating if efficacy was reached for each target parameter at that stage or at a previous stage.}
}
\value{
\link{eff.trial.any} returns a logical constant.
}
\description{
allows stopping the trial for efficacy if \emph{at least one} target parameter reached efficacy at the look of interest.
}
