% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{fut.trial.all}
\alias{fut.trial.all}
\title{trial futility stop}
\usage{
fut.trial.all(fut.target)
}
\arguments{
\item{fut.target}{logical vector with length of active intervention arms}
}
\value{
\link{fut.trial.all} returns a logical constant.
}
\description{
allows stopping the trial for futility at a given look if all active treatments reached efficacy
}
