% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_generic.R
\name{print.bats}
\alias{print.bats}
\title{Print function for BATS outputs}
\usage{
\method{print}{bats}(x, ...)
}
\arguments{
\item{x}{An object of class 'bats'.}

\item{...}{Additional arguments affecting the print produced.}
}
\description{
Print method function for objects of class 'bats'.
}
