% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.batss.R
\name{print.summary.batss}
\alias{print.summary.batss}
\title{Print the summary function for 'BATSS' outputs}
\usage{
\method{print}{summary.batss}(x, ...)
}
\arguments{
\item{x}{An object of class 'summary.batss' (i.e., output of the function \link{batss.glm}).}

\item{...}{For future use}
}
\description{
Print method function for objects of class 'summary.batss'.
}
\seealso{
\code{\link[=batss.glm]{batss.glm()}}, \code{\link[=batss.surv]{batss.surv()}}, the functions generating S3 objects of class 'batss'.
}
