% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.batss.R
\name{print.summary.batss}
\alias{print.summary.batss}
\title{Print function for objects of class 'summary.batss'}
\usage{
\method{print}{summary.batss}(x, ...)
}
\arguments{
\item{x}{An object of class 'summary.batss' (i.e., output of the function \link{summary} used on an output of the function \link{batss.glm}).}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Prints a summary for objects of class 'batss'.
}
\description{
Print function for objects of class 'summary.batss'
}
\seealso{
\code{\link[=batss.glm]{batss.glm()}}, the function generating S3 objects of class 'batss'.
}
