% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_functions.R
\name{RAR.trippa}
\alias{RAR.trippa}
\title{RAR of Trippa et al. (2012)}
\usage{
RAR.trippa(posterior, n, N, ref, active, gamma, eta, nu)
}
\arguments{
\item{posterior}{the BATS ingredient '\code{posterior}' corresponding, in this context, to the (posterior) probability P(beta>delta.RAR) for a given target parameter.}

\item{n}{the BATS ingredient '\code{n}' corresponding to the vector of number of recruited participants per arm including the control group at the look of interest.}

\item{N}{the BATS ingredient '`N' corresponding to the maximum (planned) sample size.}

\item{ref}{the BATS ingredient '\code{ref}' corresponding to a \link[base]{logical} vector of length "number of arms plus 1 (for the reference group)" and indicating which group is the reference one.}

\item{active}{the BATS ingredient '\code{active}' corresponding to a \link[base]{logical} vector of the same length as argument \code{which} (i.e., the number of target parameters) indicating if each arm is active at the look of interest.}

\item{gamma}{a scaling factor (to be defined in \code{fut.arm.control}).}

\item{eta}{a scaling factor (to be defined in \code{fut.arm.control}).}

\item{nu}{a scaling factor (to be defined in \code{fut.arm.control}).}
}
\value{
\link{RAR.trippa} returns a vector of probabilities with length of active.
}
\description{
define the group allocation probabilities based on the response adaptive randomisation rule of Trippa et al. (2012)
}
