% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.batss.R
\name{summary.batss}
\alias{summary.batss}
\title{Summary function for 'BATSS' outputs}
\usage{
\method{summary}{batss}(object, full = FALSE, ...)
}
\arguments{
\item{object}{An object of class 'batss' (i.e., output of the function \link{batss.glm}).}

\item{full}{A logical indicating if a standard (full = FALSE, default) or extended output (full = TRUE) should be returned.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Prints a summary for objects of class 'batss'.
}
\description{
Summary method function for objects of class 'batss'.
}
\seealso{
\code{\link[=batss.glm]{batss.glm()}}, the function generating S3 objects of class 'batss'.
}
