% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.batss.R
\name{summary.batss}
\alias{summary.batss}
\title{Summary function for 'BATSS' outputs}
\usage{
\method{summary}{batss}(object, extended = NULL, ...)
}
\arguments{
\item{object}{An object of class 'batss' (i.e., output of the function \link{batss.glm}).}

\item{extended}{A logical indicating if a standard (extended = FALSE, default) or extended output (extended = TRUE) should be returned. Default to \code{NULL} in which case the input of the argument \code{extended} chosen when generating \code{object} with \code{\link[=batss.glm]{batss.glm()}} is used.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Object of class 'summary.batss'.
}
\description{
Summary method function for objects of class 'batss'.
}
\seealso{
\code{\link[=batss.glm]{batss.glm()}}, the function generating S3 objects of class 'batss'.
}
